[feat] 검색 후 정렬 구현

+++ Search/src/main/java/com/rainbowgon/search/global/response/PageInfo.java
@@ -11,6 +11,6 @@ public class PageInfo {
 
     private int page;
     private int size;
-    private int totalElements;
+    private Long totalElements;
     private int totalPages;
 }
